+++ server/src/main/scala/org/http4s/server/package.scala
@@ -33,6 +33,12 @@ package object server {
    */
   type AuthMiddleware[T] = Middleware[AuthedRequest[T], Response, Request, Response]
 
+  /**
+    * Old name for SSLConfig
+    */
+  @deprecated("Use SSLConfig", "2016-12-31")
+  type SSLBits = SSLConfig
+
   object AuthMiddleware {
     def apply[T](authUser: Service[Request, T]): AuthMiddleware[T] = {
       service => service.compose(AuthedRequest(authUser))
